+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## [dev]
+- tests: start testing the cli interface
+
 ## [1.4.6] - 2019-03-30
 - remove total payments from leaderboard query (#13)
 - fix `get_staking_leaderboard`